+++ R/Spectra-functions.R
@@ -816,3 +816,17 @@ deisotopeSpectra <-
 reduceSpectra <- function(x, tolerance = 0, ppm = 10) {
     addProcessing(x, .peaks_reduce, tolerance = tolerance, ppm = ppm)
 }
+
+#' @rdname Spectra
+#'
+#' @author Johannes Rainer
+#'
+#' @export
+scalePeaks <- function(x, by = sum, msLevel. = seq_len(4)) {
+    x <- addProcessing(x, .peaks_scale_intensities, msLevel = msLevel.,
+                       by = by, spectraVariables = "msLevel")
+    x@processing <- .logging(
+        x@processing, "Scale peak intensities in spectra of MS level(s) ",
+        paste0(msLevel., collapse = ", "), ".")
+    x
+}