+++ tests/testthat/test_peaks-functions.R
@@ -345,3 +345,17 @@ test_that(".peaks_reduce works", {
     ref <- do.call(rbind, ref)
     expect_equal(ref, res)
 })
+
+test_that(".peaks_scale_intensities works", {
+    x <- cbind(mz = 1:3, intensity = c(20, 23, 14))
+    res <- .peaks_scale_intensities(x, spectrumMsLevel = 1L, msLevel = 1L)
+    expect_equal(res[, 2], x[, 2] / sum(x[, 2]))
+
+    res <- .peaks_scale_intensities(x, by = max, spectrumMsLevel = 1L,
+                                    msLevel = 1L)
+    expect_equal(res[, 2], x[, 2] / max(x[, 2]))
+
+    res <- .peaks_scale_intensities(x, by = sum, spectrumMsLevel = 1L,
+                                    msLevel = 2L)
+    expect_equal(res, x)
+})